refactor test_openapi
