+++ tests/data/expected/main_kr/main_base_class/output.py
@@ -1,0 +1,71 @@
+# generated by datamodel-codegen:
+#   filename:  api.yaml
+#   timestamp: 2019-07-26T00:00:00+00:00
+
+from __future__ import annotations
+
+from typing import List, Optional
+
+from pydantic import AnyUrl, Field
+
+from custom_module import Base
+
+
+class Pet(Base):
+    id: int
+    name: str
+    tag: Optional[str] = None
+
+
+class Pets(Base):
+    __root__: List[Pet]
+
+
+class User(Base):
+    id: int
+    name: str
+    tag: Optional[str] = None
+
+
+class Users(Base):
+    __root__: List[User]
+
+
+class Id(Base):
+    __root__: str
+
+
+class Rules(Base):
+    __root__: List[str]
+
+
+class Error(Base):
+    code: int
+    message: str
+
+
+class Api(Base):
+    apiKey: Optional[str] = Field(
+        None, description='To be used as a dataset parameter value'
+    )
+    apiVersionNumber: Optional[str] = Field(
+        None, description='To be used as a version parameter value'
+    )
+    apiUrl: Optional[AnyUrl] = Field(
+        None, description="The URL describing the dataset's fields"
+    )
+    apiDocumentationUrl: Optional[AnyUrl] = Field(
+        None, description='A URL to the API console for each API'
+    )
+
+
+class Apis(Base):
+    __root__: List[Api]
+
+
+class Event(Base):
+    name: Optional[str] = None
+
+
+class Result(Base):
+    event: Optional[Event] = None
