+++ tests/data/expected/main_kr/main_modular/foo/__init__.py
@@ -1,0 +1,20 @@
+# generated by datamodel-codegen:
+#   filename:  modular.yaml
+#   timestamp: 1985-10-26T08:21:00+00:00
+
+from __future__ import annotations
+
+from typing import Optional
+
+from pydantic import BaseModel
+
+from .. import Id
+
+
+class Tea(BaseModel):
+    flavour: Optional[str] = None
+    id: Optional[Id] = None
+
+
+class Cocoa(BaseModel):
+    quality: Optional[int] = None
