+++ tests/data/expected/main_kr/main_modular/woo/boo.py
@@ -1,0 +1,17 @@
+# generated by datamodel-codegen:
+#   filename:  modular.yaml
+#   timestamp: 1985-10-26T08:21:00+00:00
+
+from __future__ import annotations
+
+from typing import Optional
+
+from pydantic import BaseModel
+
+from .. import Source, foo
+
+
+class Chocolate(BaseModel):
+    flavour: Optional[str] = None
+    source: Optional[Source] = None
+    cocoa: Optional[foo.Cocoa] = None
