+++ tests/data/expected/openapi/openapi_parser_parse_additional_properties/with_import_format_custom_module.Base.py
@@ -1,0 +1,51 @@
+from __future__ import annotations
+
+from typing import List, Optional
+
+from pydantic import Extra
+
+from custom_module import Base
+
+
+class Pet(Base):
+    id: int
+    name: str
+    tag: Optional[str] = None
+
+
+class Pets(Base):
+    __root__: List[Pet]
+
+
+class User(Base):
+    class Config:
+        extra = Extra.allow
+
+    id: int
+    name: str
+    tag: Optional[str] = None
+
+
+class Users(Base):
+    __root__: List[User]
+
+
+class Id(Base):
+    __root__: str
+
+
+class Rules(Base):
+    __root__: List[str]
+
+
+class Error(Base):
+    code: int
+    message: str
+
+
+class Event(Base):
+    name: Optional[str] = None
+
+
+class Result(Base):
+    event: Optional[Event] = None
