+++ tests/data/expected/openapi/openapi_parser_parse_modular/__init__.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+from typing import Optional
+
+from pydantic import BaseModel
+
+from . import models
+
+
+class Id(BaseModel):
+    __root__: str
+
+
+class Error(BaseModel):
+    code: int
+    message: str
+
+
+class Result(BaseModel):
+    event: Optional[models.Event] = None
+
+
+class Source(BaseModel):
+    country: Optional[str] = None
