+++ tests/data/expected/openapi/openapi_parser_parse_modular/foo/__init__.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from typing import Optional
+
+from pydantic import BaseModel
+
+from .. import Id
+
+
+class Tea(BaseModel):
+    flavour: Optional[str] = None
+    id: Optional[Id] = None
+
+
+class Cocoa(BaseModel):
+    quality: Optional[int] = None
