+++ tests/test_main_kr.py
@@ -12,6 +12,8 @@ from datamodel_code_generator.__main__ import Exit, main
 
 DATA_PATH: Path = Path(__file__).parent / 'data'
 OPEN_API_DATA_PATH: Path = DATA_PATH / 'openapi'
+EXPECTED_MAIN_KR_PATH = DATA_PATH / 'expected' / 'main_kr'
+
 
 TIMESTAMP = '1985-10-26T01:21:00-07:00'
 
@@ -31,76 +33,7 @@ def test_main():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[Event] = None
-'''
+            == (EXPECTED_MAIN_KR_PATH / 'main' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
@@ -124,78 +57,7 @@ def test_main_base_class():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, Field
-
-from custom_module import Base
-
-
-class Pet(Base):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(Base):
-    __root__: List[Pet]
-
-
-class User(Base):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(Base):
-    __root__: List[User]
-
-
-class Id(Base):
-    __root__: str
-
-
-class Rules(Base):
-    __root__: List[str]
-
-
-class Error(Base):
-    code: int
-    message: str
-
-
-class Api(Base):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(Base):
-    __root__: List[Api]
-
-
-class Event(Base):
-    name: Optional[str] = None
-
-
-class Result(Base):
-    event: Optional[Event] = None
-'''
+            == (EXPECTED_MAIN_KR_PATH / 'main_base_class' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
@@ -219,286 +81,16 @@ def test_target_python_version():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List['Pet']
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List['User']
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List['Api']
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional['Event'] = None
-'''
+            == (
+                EXPECTED_MAIN_KR_PATH / 'target_python_version' / 'output.py'
+            ).read_text()
         )
 
     with pytest.raises(SystemExit):
         main()
 
 
-@pytest.mark.parametrize(
-    'expected',
-    [
-        {
-            (
-                '__init__.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from . import models
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Result(BaseModel):
-    event: Optional[models.Event] = None
-
-
-class Source(BaseModel):
-    country: Optional[str] = None
-''',
-            (
-                'models.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from enum import Enum
-from typing import Any, Dict, List, Optional, Union
-
-from pydantic import BaseModel
-
-
-class Species(Enum):
-    dog = 'dog'
-    cat = 'cat'
-    snake = 'snake'
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-    species: Optional[Species] = None
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Event(BaseModel):
-    name: Optional[Union[str, float, int, bool, Dict[str, Any], List[str]]] = None
-''',
-            (
-                'collections.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-from . import models
-
-
-class Pets(BaseModel):
-    __root__: List[models.Pet]
-
-
-class Users(BaseModel):
-    __root__: List[models.User]
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-''',
-            (
-                'foo',
-                '__init__.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from .. import Id
-
-
-class Tea(BaseModel):
-    flavour: Optional[str] = None
-    id: Optional[Id] = None
-
-
-class Cocoa(BaseModel):
-    quality: Optional[int] = None
-''',
-            (
-                'foo',
-                'bar.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Any, Dict, List, Optional
-
-from pydantic import BaseModel
-
-
-class Thing(BaseModel):
-    attributes: Optional[Dict[str, Any]] = None
-
-
-class Thang(BaseModel):
-    attributes: Optional[List[Dict[str, Any]]] = None
-
-
-class Clone(Thing):
-    pass
-''',
-            (
-                'woo',
-                '__init__.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-''',
-            (
-                'woo',
-                'boo.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from .. import Source, foo
-
-
-class Chocolate(BaseModel):
-    flavour: Optional[str] = None
-    source: Optional[Source] = None
-    cocoa: Optional[foo.Cocoa] = None
-''',
-        }
-    ],
-)
-def test_main_modular(
-    tmpdir_factory: TempdirFactory, expected: Mapping[str, str]
-) -> None:
+def test_main_modular(tmpdir_factory: TempdirFactory) -> None:
     """Test main function on modular file."""
 
     output_directory = Path(tmpdir_factory.mktemp('output'))
@@ -508,10 +100,10 @@ def test_main_modular(
 
     with freeze_time(TIMESTAMP):
         main(['--input', str(input_filename), '--output', str(output_path)])
-
-    for key, value in expected.items():
-        result = output_path.joinpath(*key).read_text()
-        assert result == value
+    main_modular_dir = EXPECTED_MAIN_KR_PATH / 'main_modular'
+    for path in main_modular_dir.rglob('*.py'):
+        result = output_path.joinpath(path.relative_to(main_modular_dir)).read_text()
+        assert result == path.read_text()
 
 
 def test_main_modular_no_file() -> None:
@@ -536,83 +128,7 @@ def test_main_modular_filename(tmpdir_factory: TempdirFactory) -> None:
     )
 
 
-@pytest.mark.parametrize(
-    'expected',
-    [
-        '''\
-# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[Event] = None
-'''
-    ],
-)
-def test_main_no_file(capsys: CaptureFixture, expected: str) -> None:
+def test_main_no_file(capsys: CaptureFixture) -> None:
     """Test main function on non-modular file with no output name."""
 
     input_filename = OPEN_API_DATA_PATH / 'api.yaml'
@@ -621,79 +137,15 @@ def test_main_no_file(capsys: CaptureFixture, expected: str) -> None:
         main(['--input', str(input_filename)])
 
     captured = capsys.readouterr()
-    assert captured.out == expected
+    assert (
+        captured.out
+        == (EXPECTED_MAIN_KR_PATH / 'main_no_file' / 'output.py').read_text()
+    )
+
     assert not captured.err
 
 
-@pytest.mark.parametrize(
-    'expected',
-    [
-        '''\
-# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):  # 1 2, 1 2, this is just a pet
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = None
-    apiVersionNumber: Optional[str] = None
-    apiUrl: Optional[AnyUrl] = None
-    apiDocumentationUrl: Optional[AnyUrl] = None
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[Event] = None
-'''
-    ],
-)
-def test_main_custom_template_dir(capsys: CaptureFixture, expected: str) -> None:
+def test_main_custom_template_dir(capsys: CaptureFixture) -> None:
     """Test main function with custom template directory."""
 
     input_filename = OPEN_API_DATA_PATH / 'api.yaml'
@@ -713,7 +165,12 @@ def test_main_custom_template_dir(capsys: CaptureFixture, expected: str) -> None
         )
 
     captured = capsys.readouterr()
-    assert captured.out == expected
+    assert (
+        captured.out
+        == (
+            EXPECTED_MAIN_KR_PATH / 'main_custom_template_dir' / 'output.py'
+        ).read_text()
+    )
     assert not captured.err
 
 
@@ -735,99 +192,7 @@ def test_pyproject():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import (
-    annotations,
-)
-
-from typing import (
-    List,
-    Optional,
-)
-
-from pydantic import (
-    AnyUrl,
-    BaseModel,
-    Field,
-)
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[
-        str
-    ] = Field(
-        None,
-        description="To be used as a dataset parameter value",
-    )
-    apiVersionNumber: Optional[
-        str
-    ] = Field(
-        None,
-        description="To be used as a version parameter value",
-    )
-    apiUrl: Optional[
-        AnyUrl
-    ] = Field(
-        None,
-        description="The URL describing the dataset\'s fields",
-    )
-    apiDocumentationUrl: Optional[
-        AnyUrl
-    ] = Field(
-        None,
-        description="A URL to the API console for each API",
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[
-        Event
-    ] = None
-'''
+            == (EXPECTED_MAIN_KR_PATH / 'pyproject' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
